+++ pkg/agent/catalog/svidstore.go
@@ -2,7 +2,7 @@ package catalog
 
 import (
 	"github.com/spiffe/spire/pkg/agent/plugin/svidstore"
-	"github.com/spiffe/spire/pkg/agent/plugin/svidstore/aws"
+	"github.com/spiffe/spire/pkg/agent/plugin/svidstore/awssecretsmanager"
 	"github.com/spiffe/spire/pkg/common/catalog"
 )
 
@@ -24,7 +24,7 @@ func (repo *svidStoreRepository) Versions() []catalog.Version {
 
 func (repo *svidStoreRepository) BuiltIns() []catalog.BuiltIn {
 	return []catalog.BuiltIn{
-		aws.BuiltIn(),
+		awssecretsmanager.BuiltIn(),
 	}
 }
