+++ pkg/agent/plugin/svidstore/awssecretsmanager/aws_test.go
@@ -1,4 +1,4 @@
-package aws
+package awssecretsmanager
 
 import (
 	"context"
