+++ pkg/agent/plugin/svidstore/awssecretsmanager/client.go
@@ -1,4 +1,4 @@
-package aws
+package awssecretsmanager
 
 import (
 	"context"
